Create checkSeqListMismatchQiime.py
This script is for troubleshooting issues in Qiime2. It checks for mismatches between feature tables (ASV-table) and sequence lists (rep-seq) to make sure that everything in the feature table is also present in rep-seq. It prints any feature IDs that are present in the feature table but not in rep-seq.